[NO-ISSUE] Fixed issue with FetchingProvider remounting children (#14)
* There was an issue with FetchingProvider that it rendered it's
children initially, before fetching was started, then rendering null and
finally rendering children again after fetching. This happened when
blocking was set to true (which is default). This was causing multiple
mounts of children, affecting performance.
* The issue was caused by misuse of "isFetching" flag in rendering logic
since it was initially false, before fetching has even started.
* Replacing "isFetching" with "isFetched" is restoring original purpose
of "blocking" flag, since rendering should or shouldn't be blocked until
data is fetched, not while it is fetching as it was implemented so far.